implement only necessary types
