--- src/request/types/permissions/utils/index.ts
@@ -1,7 +1,0 @@
-import { makeAccountId, accountIdBrandName, accountIdSchema } from './account-id';
-export const account = {
-  makeAccountId,
-  accountIdBrandName,
-  accountIdSchema,
-};
-export type * as Account from './account-id';
